Add LOTS of error correction at spreadsheet, fix ability to insert into DB
A LOT of work went in to get to this point. I tried many different ways of trying
to handle leading zeros in DATE of xlsx spreadsheet that was throwing off format.
Including trying to correct it after having been read by phpspreadsheet tool by
iterating over every cell and trimming it. I tried a utf8 decode followed by a trim and then
a re-encode.

In the end, I had to correct it from the date string in the csv. This involved removing
the outside single quotes, then trimming it, then tokenizing the string with "/" as a
delimiter, and then checking if the token was length > 2. If so, it was the year (4 digits).
The day and month always come in that respective order, so at that point it was just a
matter of tokenizing the string two more times, and then changing the format to match
yyyy-dd-mm.

At this point, the code trims any leading whitespace or whitespace-like characters,
replaces any multiple-spaces that were not on the outside of the string with a single
whitespace (This was throwing off my Ability to successfully execute mysql queries).

Feeling pretty accomplished at the moment :-)     A lot of work went into this.
This is probably evidenced by all the commented out code that I have yet to remove.